correction diverse

+++ ft_itoa.c
@@ -45,7 +45,7 @@ static char	*filling(int n, int x, char *converted)
 char	*ft_itoa(int n)
 {
 	size_t		x;
-	size_t		nb;
+	int			nb;
 	char		*converted;
 
 	if ((n == (-2147483647 -1)) || n == 0)
